backend/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.4.4
Werkzeug==2.3.7
python-dotenv==1.0.1
cryptography==41.0.3
PyJWT==2.8.0
flask-cors==3.0.10
bcrypt==4.0.1
Flask-Migrate==4.0.4
SQLAlchemy==2.0.20
Flask-Mail==0.10.0
APScheduler==3.9.1
